Rename commits file
